moved constants into zetaclient
