+++ zetaclient/common/utils_test.go
@@ -51,3 +51,78 @@ func TestCctxRestricted(t *testing.T) {
 		require.False(t, IsCctxRestricted(&cctx))
 	})
 }
+
+func Test_GasPriceMultiplier(t *testing.T) {
+	tt := []struct {
+		name       string
+		chainID    int64
+		multiplier float64
+		fail       bool
+	}{
+		{
+			name:       "get Ethereum multiplier",
+			chainID:    1,
+			multiplier: 1.2,
+			fail:       false,
+		},
+		{
+			name:       "get Goerli multiplier",
+			chainID:    5,
+			multiplier: 1.2,
+			fail:       false,
+		},
+		{
+			name:       "get BSC multiplier",
+			chainID:    56,
+			multiplier: 1.2,
+			fail:       false,
+		},
+		{
+			name:       "get BSC Testnet multiplier",
+			chainID:    97,
+			multiplier: 1.2,
+			fail:       false,
+		},
+		{
+			name:       "get Polygon multiplier",
+			chainID:    137,
+			multiplier: 1.2,
+			fail:       false,
+		},
+		{
+			name:       "get Mumbai Testnet multiplier",
+			chainID:    80001,
+			multiplier: 1.2,
+			fail:       false,
+		},
+		{
+			name:       "get Bitcoin multiplier",
+			chainID:    8332,
+			multiplier: 2.0,
+			fail:       false,
+		},
+		{
+			name:       "get Bitcoin Testnet multiplier",
+			chainID:    18332,
+			multiplier: 2.0,
+			fail:       false,
+		},
+		{
+			name:       "get unknown chain gas price multiplier",
+			chainID:    1234,
+			multiplier: 1.0,
+			fail:       true,
+		},
+	}
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			multiplier, err := GasPriceMultiplier(tc.chainID)
+			if tc.fail {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			require.Equal(t, tc.multiplier, multiplier)
+		})
+	}
+}
